requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.0
django-cors-headers>=4.0
psycopg2-binary>=2.9
django-redis>=5.3
redis>=5.0
channels[daphne]>=4.0
uvicorn[standard]>=0.23
gunicorn>=21.2
celery>=5.3
django-celery-beat>=2.5
Pillow>=10.0
python-decouple>=3.8
razorpay>=1.4
boto3>=1.28
django-storages>=1.14
requests>=2.31
django-ipware>=5.0
sentry-sdk>=1.30
drf-spectacular>=0.26
django-prometheus>=2.3
python-magic==0.4.27
Faker>=19.0.0
geopy>=2.4
django-leaflet>=0.29.0

dj-database-url
psycopg2-binary
gunicorn
whitenoise

dj-database-url